Restructured project for better modularity
Originally all of the functionality was kept in `notify.py`. This was
cumbersome and hard to follow. In an effort to make the project more
modular and easier to contribute to `notify.py` has been replaced with
`main.py`. The bulk of the code has been moved inside a `modules`
directory and imported to `main.py` with some simple logic wrapped
around the necessary function.

Signed-off-by: discr33t <cpisano86@gmail.com>