Implement Recently Added TV and Refactor
The work to implement retreiving the recently added tv required more
refactors than originally inteded. Along with this work changes were
also made to the Travis config as well as local testing config to
implement Tusk. A change to the module structure and how modules were
imported to main.py was done. Separate modules for getting movies and
media were created as well as a module for utilities.

Signed-off-by: discr33t <cpisano86@gmail.com>