Refactor plex_movies (#27)
plex_movies.py is refactored to return a nested dictionary that includes
title, summary, and year. main.py is refactored to create a new nested
dictionary consisting of the plex_movies.py returned dictionary and the
plex_tv.py returned dictionary. The aggregated dictionary is printed to
the console as a json string.

*email needs to be completely refactored

Signed-off-by: discr33t <cpisano86@gmail.com>