v0.2.0 Release (#34)
* Implement Recently Added TV and Refactor (#26)

The work to implement retreiving the recently added tv required more
refactors than originally inteded. Along with this work changes were
also made to the Travis config as well as local testing config to
implement Tusk. A change to the module structure and how modules were
imported to main.py was done. Separate modules for getting movies and
media were created as well as a module for utilities.

Signed-off-by: discr33t <cpisano86@gmail.com>

* Refactor plex_movies (#27)

plex_movies.py is refactored to return a nested dictionary that includes
title, summary, and year. main.py is refactored to create a new nested
dictionary consisting of the plex_movies.py returned dictionary and the
plex_tv.py returned dictionary. The aggregated dictionary is printed to
the console as a json string.

*email needs to be completely refactored

Signed-off-by: discr33t <cpisano86@gmail.com>

* Implemented Jinja Template and Python 3 (#29) (#31)

Switched to Python 3, Pip3 utilizing pipenv to manage dependencies.
Updated Tusk and Travis accordingly. Modified email to use a Jinja
template to format the body and changed to a better email library.

Signed-off-by: discr33t <cpisano86@gmail.com>

* Test Jinja Template (#32)

Replaced the removed test validating the formatting of the email body
and replaced it with a test that used mock data to validate the
rendering of the Jinja template.

Signed-off-by: discr33t <cpisano86@gmail.com>

* Update Changelog for Release (#33)

Signed-off-by: discr33t <cpisano86@gmail.com>